requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-cursor-pagination==0.2.1
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
ghp-import==2.1.0
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.3.7
Markdown==3.6
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.46
psycopg==3.1.19
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyyaml_env_tag==0.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
vine==5.1.0
watchdog==4.0.1
wcwidth==0.2.13
